fix bad timestamp for API print event (#140)
* prevent latest timestamp from going back in time

It was possible for an analog message to enter the message processing queue and have its older timestamp assigned as the latest timestamp. If the timing was just right, this could occasionally cause the get_timestamp() function to produce an out of sequence/incorrect timestamp.

This commit fixes the problem by only updating the latest timestamp with the message's timestamp, if the message's timestamp is newer.